refactor: move CameraExtension class to its own file (#1064)
